+++ lib/logger/utils.ts
@@ -339,3 +339,16 @@ export function getEnv(key: string): string | undefined {
     .map((v) => v?.toLowerCase().trim())
     .find(is.nonEmptyStringAndNotWhitespace);
 }
+
+export function getMessage(
+  p1: string | Record<string, any>,
+  p2?: string,
+): string | undefined {
+  return is.string(p1) ? p1 : p2;
+}
+
+export function toMeta(
+  p1: string | Record<string, any>,
+): Record<string, unknown> {
+  return is.object(p1) ? p1 : {};
+}
